+++ website/src/layout.rs
@@ -6,24 +6,55 @@ use leptos_bulma::components::{
 };
 use leptos_bulma::elements::{BButton, BButtons, BIcon};
 use leptos_bulma::icons::icondata_fa;
+use leptos_use::ColorMode;
 
-use crate::app::use_theme_context;
+use crate::app::use_app_color_mode;
 use crate::i18n::{t, use_i18n, Locale};
+
+#[component]
+fn ImageColorModes(
+    dark_src: &'static str,
+    light_src: &'static str,
+    alt: &'static str,
+    width: i8,
+) -> impl IntoView {
+    let color_mode = use_app_color_mode().mode;
+
+    view! {
+        <picture>
+            <Show when=move || [ColorMode::Dark, ColorMode::Auto].contains(&color_mode.get())>
+                <source srcset=dark_src media="(prefers-color-scheme: dark)"/>
+            </Show>
+
+            <Show when=move || [ColorMode::Light, ColorMode::Auto].contains(&color_mode.get())>
+                <source srcset=light_src media="(prefers-color-scheme: light)"/>
+            </Show>
+
+            <img
+                src=move || { if color_mode.get() == ColorMode::Dark { dark_src } else { light_src } }
+                alt=alt
+                width=width
+            />
+        </picture>
+    }
+}
 
 #[component]
 pub fn Layout(children: Children) -> impl IntoView {
     let i18n = use_i18n();
-    let theme = use_theme_context();
-    let theme_is_dark = create_rw_signal(theme.is_dark());
-    let theme_is_light = create_rw_signal(theme.is_light());
-    let theme_is_system = create_rw_signal(theme.is_system());
+    let app_color_mode = use_app_color_mode();
+    let color_mode = app_color_mode.mode.get();
+    let color_mode_is_dark = create_rw_signal(color_mode == ColorMode::Dark);
+    let color_mode_is_light = create_rw_signal(color_mode == ColorMode::Light);
+    let color_mode_is_auto = create_rw_signal(color_mode == ColorMode::Auto);
     let burger_is_active = create_rw_signal(false);
 
     create_effect(move |_| {
-        theme.get();
-        theme_is_dark.set(theme.is_dark());
-        theme_is_light.set(theme.is_light());
-        theme_is_system.set(theme.is_system());
+        let color_mode = app_color_mode.mode.get();
+
+        color_mode_is_dark.set(color_mode == ColorMode::Dark);
+        color_mode_is_light.set(color_mode == ColorMode::Light);
+        color_mode_is_auto.set(color_mode == ColorMode::Auto);
     });
 
     view! {
@@ -78,18 +109,19 @@ pub fn Layout(children: Children) -> impl IntoView {
                         <div class="is-flex is-align-items-center is-justify-content-center">
                             {t!(i18n, this_website_was_made_with)}
                             <a class="mx-3" href="https://leptos.dev" target="_blank" title="Go to Leptos">
-                                <picture>
-                                    <source
-                                        srcset="/images/leptos-logo-light.svg"
-                                        media="(prefers-color-scheme: dark)"
-                                    />
-                                    <img src="/images/leptos-logo.svg" alt="Leptos" width="100"/>
-                                </picture>
+                                <ImageColorModes
+                                    dark_src="/images/leptos-logo-light.svg"
+                                    light_src="/images/leptos-logo.svg"
+                                    alt="Leptos"
+                                    width=100
+                                />
                             </a> & <a class="mx-3" href="https://bulma.io/" target="_blank" title="Go to Bulma">
-                                <picture>
-                                    <source srcset="/images/bulma-logo-light.svg" media="(prefers-color-scheme: dark)"/>
-                                    <img src="/images/bulma-logo.svg" alt="Bulma" width="100"/>
-                                </picture>
+                                <ImageColorModes
+                                    dark_src="/images/bulma-logo-light.svg"
+                                    light_src="/images/bulma-logo.svg"
+                                    alt="Bulma"
+                                    width=100
+                                />
                             </a>
                         </div>
                         <div class="mt-3 is-flex is-align-items-center is-justify-content-center">
@@ -100,18 +132,17 @@ pub fn Layout(children: Children) -> impl IntoView {
                                 target="_blank"
                                 title="Go to GitHub"
                             >
-                                <picture>
-                                    <source
-                                        srcset="/images/github-logo-light.svg"
-                                        media="(prefers-color-scheme: dark)"
-                                    />
-                                    <img src="/images/github-logo.svg" alt="GitHub" width="100"/>
-                                </picture>
+                                <ImageColorModes
+                                    dark_src="/images/github-logo-light.svg"
+                                    light_src="/images/github-logo.svg"
+                                    alt="GitHub"
+                                    width=100
+                                />
                             </a>
                         </div>
                     </BColumn>
 
-                    <BColumn is="narrow">
+                    <BColumn is="narrow has-text-right">
                         <BDropdown
                             class="mb-4"
                             is_right=true
@@ -129,15 +160,23 @@ pub fn Layout(children: Children) -> impl IntoView {
                             <BButton
                                 class="ml-auto"
                                 title="System theme"
-                                is_active=theme_is_system
-                                on:click=move |_| theme.set_system()
+                                is_active=color_mode_is_auto
+                                on:click=move |_| app_color_mode.set_mode.set(ColorMode::Auto)
                             >
                                 <BIcon is_scaled=false icon=icondata_fa::FaDesktopSolid/>
                             </BButton>
-                            <BButton title="Light theme" is_active=theme_is_light on:click=move |_| theme.set_light()>
+                            <BButton
+                                title="Light theme"
+                                is_active=color_mode_is_light
+                                on:click=move |_| app_color_mode.set_mode.set(ColorMode::Light)
+                            >
                                 <BIcon is_scaled=false icon=icondata_fa::FaSunSolid/>
                             </BButton>
-                            <BButton title="Dark theme" is_active=theme_is_dark on:click=move |_| theme.set_dark()>
+                            <BButton
+                                title="Dark theme"
+                                is_active=color_mode_is_dark
+                                on:click=move |_| app_color_mode.set_mode.set(ColorMode::Dark)
+                            >
                                 <BIcon is_scaled=false icon=icondata_fa::FaMoonSolid/>
                             </BButton>
                         </BButtons>