+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- Upgrade `leptos-use` in the website to include Spin feature for `use_cookie`.
+- Replace theme context in the website with `leptos_use::use_color_mode`.
+
+### Fixed
+
+- Toggle footer images in the website with light and dark modes.
+
+### Removed
+
+- Custom `use_cookie` for SSR in the website.
+
 ## [0.0.0-alpha.12] - 2024-04-17
 
 ### Added